Adding changes made in friday

+++ src/main/java/eu/deltasorce/internship/library/controller/AuthorController.java
@@ -1,40 +1,31 @@
 package eu.deltasorce.internship.library.controller;
 
 import eu.deltasorce.internship.library.model.book.Author;
+import eu.deltasorce.internship.library.repository.AuthorRepository;
 import eu.deltasorce.internship.library.service.AuthorService;
 
-import java.time.LocalDate;
-
-/**
- * Class who is responsible for the commands
- */
 public class AuthorController {
-
-    private Author author;
 
     private final AuthorService service = new AuthorService();
 
-    public boolean add(String name, String country, LocalDate birthDate) {
-        author = new Author(name, country, birthDate);
+    public boolean add(Author author) {
+        if (author == null) {
+            throw new IllegalArgumentException("Author was not given");
+        }
         return service.add(author);
     }
 
-    public boolean add(String name, String country, LocalDate birthDate, LocalDate deathDate) {
-        author = new Author(name, country, birthDate, deathDate);
-        return service.add(author);
-    }
-
-    public boolean delete(String name, String country, LocalDate birthDate) {
-        author = new Author(name, country, birthDate);
+    public boolean delete(Author author) {
+        if (author == null) {
+            throw new IllegalArgumentException("Author was not given");
+        }
         return service.delete(author);
     }
 
-    public boolean delete(String name, String country, LocalDate birthDate, LocalDate deathDate) {
-        author = new Author(name, country, birthDate, deathDate);
-        return service.delete(author);
-    }
-
-    public boolean deleteById(int index) {
-        return service.deleteById(index);
+    public boolean deleteById(int id) {
+        if (id < 0 ) {
+            throw new IllegalArgumentException("ID cant be negative");
+        }
+        return service.deleteById(id);
     }
 }
